CSN: CSN-platform-acp-7
Completed implementation of Arches MPI Device and a new Physical Channel
based on the device. An RRRTest model can now be built for the ACP
Platform, but a slight hacking of the SCons script is needed to link
the two libraries libacp.a and libmpi_xxx.a correctly into the host
software executable.

The platform currently does not have a clock divider/multiplier, so only
simple models that can run at 133MHz will work.