CSN: CSN-platform-acp-14
Moved MAKE_ALL_TARGET from the root modules (model/project) to the
physical_platform module inside fpgaenv. In the various physical
platforms the default is set appropriately, e.g., exe for simulation,
bit for real hardware. In the physical_platform module, MAKE_ALL_TARGET
is declared as a 'global' parameter so it can be overridden in the
model that requires an fpgaenv.