CSN: CSN-platform-acp-76
The check in is a waypoint along the line to having a multiple FPGA
build.  Here, I have added the nallatech provided communications
primitives and adjusted the existing edge device to support the
interface required for these devices.  I also refactored the edge
device to share as much code as possible between the real edge device
(FPGA0) and the null edge device (FPGA1).  I also added Bluespec
wrappers for the intra ACP communications primitives.  


The next step will be to add a physical platform for FPGA1 and a new
platform with the communication stub for FPGA0.