CSN: CSN-platform-acp-78
The main addition in this large check-in is a functional test for the
intra-fpga lvds channel.  I also managed to get control of the
on-board leds on the ACP module.  This required setting a marginally
undocumented register.  I also attempted to get the JTAG operational,
but I was uncessful on the ACP.