CSN: CSN-leap-platforms-acp-120
During initialization of a pair of FPGAs within a module, FPGA0 is allowed to
write to FPGA1 in spite of it not being ready.  Added a simple initialization
protocol, where FPGA0 waits to hear from FPGA1 before sending any messages.
I can't find any Nallatech documentation on a better method.  This simple
protocol probably depends on the order in which each FPGA is initialized,
though I didn't test that.  Also added comments to the C++ initialization
code, noting the fragility of the startup sequence.